+++ R/game_systems.R
@@ -32,6 +32,11 @@
 #' \item{hexpack}{A hexagonal extrapolation of the piecepack designed by Nathan Morse and Daniel Wilcox.
 #'                See \url{https://boardgamegeek.com/boardgameexpansion/35424/hexpack}.}
 #' \item{meeples}{Standard 16mm x 16mm x 10mm \dQuote{meeples} in six colors represented by a \dQuote{bit}.}
+#' \item{morris}{Various morris aka mills aka merels games in six colors.
+#'               Color is controlled by suit and \dQuote{size} of morris board
+#'               is controlled by rank e.g. \dQuote{Six men's morris} corresponds to a rank of 6 and
+#'               \dQuote{Nine men's morris} corresponds to a rank of 9.
+#'               Game pieces are represented by stones.}
 #' \item{piecepack}{A public domain game system invented by James "Kyle" Droscha.
 #'   See \url{https://www.ludism.org/ppwiki}.
 #'   Configuration also contains the following piecepack accessories:\describe{
@@ -136,6 +141,7 @@ game_systems <- function(style = NULL, round = FALSE, pawn = "token") {
          go = go(1, color_list),
          hexpack = packs$hexpack,
          meeples = meeples(color_list),
+         morris = morris(1, color_list),
          piecepack = packs$base,
          playing_cards = cards$base,
          playing_cards_colored = cards$color,
@@ -189,6 +195,47 @@ meeples <- function(color_list) {
     meeples
 }
 
+morris <- function(cell_width = 1, color_list) {
+    morris <- list(n_suits = 6, n_ranks = 15,
+                   width.board = 7 * cell_width,
+                   height.board = 7 * cell_width,
+                   grob_fn.r1.board_face = morrisBoardGrobFn(12),
+                   gridline_color.board_face = cb_suit_colors_pure,
+                   gridline_color.board_back = "transparent",
+                   gridline_lex.board = 4,
+                   suit_color = cb_suit_colors_impure,
+                   background_color = color_list$background_color,
+                   gridline_color.s6.board_face = "grey80")
+    for (i in seq(2, 15)) {
+        if (i < 5) {
+            morris[[paste0("width.r", i, ".board")]] <- 3 * cell_width
+            morris[[paste0("height.r", i, ".board")]] <- 3 * cell_width
+        } else if (i < 8) {
+            morris[[paste0("width.r", i, ".board")]] <- 5 * cell_width
+            morris[[paste0("height.r", i, ".board")]] <- 5 * cell_width
+        } else {
+            morris[[paste0("width.r", i, ".board")]] <- 7 * cell_width
+            morris[[paste0("height.r", i, ".board")]] <- 7 * cell_width
+        }
+        morris[[paste0("grob_fn.r", i, ".board_face")]] <- morrisBoardGrobFn(i)
+    }
+    morris <- pp_cfg(c(morris, go_stone(cell_width), color_list))
+    morris$has_piecepack <- FALSE
+    morris$has_boards <- TRUE
+    morris$has_bits <- TRUE
+    morris
+}
+
+go_stone <- function(cell_width) {
+    list(width.bit = 0.8700787 * cell_width,
+         depth.bit = 0.3937008 * cell_width,
+         invert_colors.bit = TRUE,
+         op_grob_fn.bit = basicEllipsoid,
+         obj_fn.bit = function(...) save_ellipsoid_obj(..., subdivide=4),
+         ps_text.bit = "", dm_text.bit = ""
+    )
+}
+
 dominoes <- function(background_color = "white", suit_color = "black", border_color = "black",
                      rect_shape, mat_width = 0) {
     border_lex <- ifelse(border_color == "black", 4, 0)
@@ -215,12 +262,16 @@ dominoes <- function(background_color = "white", suit_color = "black", border_co
     dominoes
 }
 
+checker_piece <- function(cell_width) {
+    list(width.bit = 0.75 * cell_width, invert_colors.bit = TRUE,
+         ps_text.bit = "", dm_text.bit = "")
+
+}
+
 checkers <- function(cell_width = 1, color_list) {
     checkers <- list(n_suits = 6, n_ranks = 12,
                      width.board = 8 * cell_width,
                      height.board = 8 * cell_width,
-                     width.bit = 0.75 * cell_width, invert_colors.bit = TRUE,
-                     ps_text.bit = "", dm_text.bit = "",
                      grob_fn.r1.board_face = checkeredBoardGrobFn(8, 8),
                      grob_fn.r1.board_back = linedBoardGrobFn(8, 8),
                      gridline_color.board_face = cb_suit_colors_impure,
@@ -236,7 +287,7 @@ checkers <- function(cell_width = 1, color_list) {
         checkers[[paste0("grob_fn.r", i, ".board_face")]] <- checkeredBoardGrobFn(i, i)
         checkers[[paste0("grob_fn.r", i, ".board_back")]] <- linedBoardGrobFn(i, i)
     }
-    checkers <- pp_cfg(c(checkers, color_list))
+    checkers <- pp_cfg(c(checkers, checker_piece(cell_width), color_list))
     checkers$has_piecepack <- FALSE
     checkers$has_boards <- TRUE
     checkers$has_bits <- TRUE
@@ -289,13 +340,7 @@ go <- function(cell_width = 1, color_list) {
     go <- list(n_suits = 6, n_ranks = 19,
                width.board = (18 + 1) * cell_width,
                height.board = (18 + 1) * cell_width,
-               width.bit = 0.8700787 * cell_width, # white stone should be 22.1 mm wide
-               depth.bit = 0.3937008 * cell_width, # stones should be 6-10 mm thick
                grob_fn.board_back = basicPieceGrob,
-               invert_colors.bit = TRUE,
-               op_grob_fn.bit = basicEllipsoid,
-               obj_fn.bit = function(...) save_ellipsoid_obj(..., subdivide=4),
-               ps_text.bit = "", dm_text.bit = "",
                grob_fn.r1.board_face = linedBoardGrobFn(18, 18, 0.5),
                gridline_color.board_face = cb_suit_colors_pure,
                gridline_color.board_back = "transparent",
@@ -308,7 +353,7 @@ go <- function(cell_width = 1, color_list) {
         go[[paste0("height.r", i, ".board")]] <- i * cell_width
         go[[paste0("grob_fn.r", i, ".board_face")]] <- linedBoardGrobFn(i - 1, i - 1, 0.5)
     }
-    go <- pp_cfg(c(go, color_list))
+    go <- pp_cfg(c(go, go_stone(cell_width), color_list))
     go$has_piecepack <- FALSE
     go$has_boards <- TRUE
     go$has_bits <- TRUE