+++ R/utils-boards.R
@@ -75,3 +75,122 @@ cycle_elements <- function(x, n = 1) {
         cycle_elements(cycle_elements(x, l), n-l)
     }
 }
+
+#### make.Content
+morrisBoardGrobFn <- function(n_pieces = 12) {
+    force(n_pieces)
+    function(piece_side, suit, rank, cfg=pp_cfg()) {
+        cfg <- as_pp_cfg(cfg)
+        opt <- cfg$get_piece_opt(piece_side, suit, rank)
+        shape <- pp_shape(opt$shape, opt$shape_t, opt$shape_r, opt$back)
+
+        background_grob <- shape$shape(gp=gpar(col=NA, fill=opt$background_color), name = "background")
+
+        gp_line <- gpar(col=NA, fill=opt$gridline_color)
+        if (n_pieces < 3) {
+            inner_grob <- morrisBoard3Grob(gp=gp_line, name = "lines_and_dots", diag = FALSE)
+        } else if (n_pieces < 5) {
+            inner_grob <- morrisBoard3Grob(gp=gp_line, name = "lines_and_dots")
+        } else if (n_pieces < 7) {
+            inner_grob <- morrisBoard6Grob(gp=gp_line, name = "lines_and_dots")
+        } else if (n_pieces == 7) {
+            inner_grob <- morrisBoard6Grob(gp=gp_line, name = "lines_and_dots", cross = TRUE)
+        } else if (n_pieces < 11) {
+            inner_grob <- morrisBoard9Grob(gp=gp_line, name = "lines_and_dots")
+        } else {
+            inner_grob <- morrisBoard9Grob(gp=gp_line, name = "lines_and_dots", diag = TRUE)
+        }
+
+        gp_border <- gpar(col=opt$border_color, fill=NA, lex=opt$border_lex)
+        border_grob <- shape$shape(gp=gpar(col=opt$border_color, fill=NA, lex=opt$border_lex), name = "border")
+
+        grobTree(background_grob, inner_grob, border_grob,
+                 name=piece_side, cl = "morris_board")
+
+    }
+}
+
+morrisBoard3Grob <- function(gp = gpar(), name = NULL, diag = TRUE) {
+    w <- 1/64
+    wd <- w / sqrt(2) / 2
+    vlines <- rectGrob(x = c(0, 1, 2) / 2, y = 0.5, width = w, height = 1 + w)
+    hlines <- rectGrob(x = 0.5, y = c(0, 1, 2) / 2, height = w, width = 1 + w)
+    if (diag) {
+        tl_br <- polygonGrob(x = c(1 - wd, 1 + wd, 1 + wd, 0 + wd, 0 - wd, 0 - wd, 1 - wd),
+                             y = c(0 - wd, 0 - wd, 0 + wd, 1 + wd, 1 + wd, 1 - wd, 0 - wd))
+        tr_bl <- polygonGrob(x = c(0 + wd, 0 - wd, 0 - wd, 1 - wd, 1 + wd, 1 + wd, 0 + wd),
+                             y = c(0 - wd, 0 - wd, 0 + wd, 1 + wd, 1 + wd, 1 - wd, 0 - wd))
+    } else {
+        tl_br <- NULL
+        tr_bl <- NULL
+    }
+    dots <- circleGrob(x = rep(c(0, 1/2, 1), 3), y = rep(c(0, 1/2, 1), each = 3), r = 1.5 * w)
+
+    grobTree(vlines, hlines, tl_br, tr_bl, dots,
+             name = name, cl = "morris_board_inner", gp = gp,
+             vp = viewport(width = 2/3, height = 2/3))
+}
+morrisBoard6Grob <- function(gp = gpar(), name = NULL, cross = FALSE) {
+    w <- 1/64
+    wd <- w / sqrt(2) / 2
+    dots <- circleGrob(x = rep.int(c(0, 1, 2, 3, 4), c(3, 3, 4, 3, 3)) / 4,
+                       y = c(0, 2, 4, 1, 2, 3, 0, 1, 3, 4, 1, 2, 3, 0, 2, 4) / 4,
+                       r = 1.5 * w)
+    if (cross) {
+        vlines <- rectGrob(x = c(0, 1, 2,  3, 4) / 4,
+                           y = c(2, 2, 2, 2, 2) / 4,
+                           width = w,
+                           height = c(4, 2, 4, 2, 4) / 4 + w)
+        hlines <- rectGrob(x =  c(2, 2, 2, 2, 2) / 4,
+                           y =  c(0, 1, 2, 3, 4) / 4,
+                           height = w,
+                           width = c(4, 2, 4, 2, 4) / 4 + w)
+    } else {
+        vlines <- rectGrob(x = c(0, 1, 2, 2, 3, 4) / 4,
+                           y = c(2, 2, 0.5, 3.5, 2, 2) / 4,
+                           width = w,
+                           height = c(4, 2, 1, 1, 2, 4) / 4 + w)
+        hlines <- rectGrob(x =  c(2, 2, 0.5, 3.5, 2, 2) / 4,
+                           y =  c(0, 1, 2, 2, 3, 4) / 4,
+                           height = w,
+                           width = c(4, 2, 1, 1, 2, 4) / 4 + w)
+    }
+
+    grobTree(vlines, hlines, dots,
+             name = name, cl = "morris_board_inner", gp = gp,
+             vp = viewport(width = 4/5, height = 4/5))
+}
+morrisBoard9Grob <- function(gp = gpar(), name = NULL, diag = FALSE) {
+    w <- 1/64
+    wd <- w / sqrt(2) / 2
+    vlines <- rectGrob(x = c(0, 1, 2, 3, 3, 4, 5, 6) / 6,
+                       y = c(3, 3, 3, 1, 5, 3, 3, 3) / 6,
+                       width = w,
+                       height = c(6, 4, 2, 2, 2, 2, 4, 6) / 6 + w)
+    hlines <- rectGrob(x =  c(3, 3, 3, 1, 5, 3, 3, 3) / 6,
+                       y =  c(0, 1, 2, 3, 3, 4, 5, 6) / 6,
+                       height = w,
+                       width =  c(6, 4, 2, 2, 2, 2, 4, 6) / 6 + w)
+    dots <- circleGrob(x = rep.int(c(0, 1, 2, 3, 4, 5, 6), c(3, 3, 3, 6, 3, 3, 3)) / 6,
+                       y = c(0, 3, 6, 1, 3, 5, 2, 3, 4, 0, 1, 2, 4, 5, 6, 2, 3, 4, 1, 3, 5, 0, 3, 6) / 6,
+                       r = 1.5 * w)
+    if (diag) {
+        ul <- polygonGrob(x = c(2/6 - wd, 2/6 + wd, 2/6 + wd, 0 + wd, 0 - wd, 0 - wd, 2/6 - wd),
+                          y = c(0 - wd, 0 - wd, 0 + wd, 2/6 + wd, 2/6 + wd, 2/6 - wd, 0 - wd) + 4/6)
+        br <- polygonGrob(x = c(2/6 - wd, 2/6 + wd, 2/6 + wd, 0 + wd, 0 - wd, 0 - wd, 2/6 - wd) + 4/6,
+                          y = c(0 - wd, 0 - wd, 0 + wd, 2/6 + wd, 2/6 + wd, 2/6 - wd, 0 - wd))
+        ur <- polygonGrob(x = c(0 + wd, 0 - wd, 0 - wd, 2/6 - wd, 2/6 + wd, 2/6 + wd, 0 + wd) + 4/6,
+                          y = c(0 - wd, 0 - wd, 0 + wd, 2/6 + wd, 2/6 + wd, 2/6 - wd, 0 - wd) + 4/6)
+        bl <- polygonGrob(x = c(0 + wd, 0 - wd, 0 - wd, 2/6 - wd, 2/6 + wd, 2/6 + wd, 0 + wd),
+                          y = c(0 - wd, 0 - wd, 0 + wd, 2/6 + wd, 2/6 + wd, 2/6 - wd, 0 - wd))
+    } else {
+        ul <- NULL
+        ur <- NULL
+        br <- NULL
+        bl <- NULL
+    }
+
+    grobTree(vlines, hlines, dots, ul, ur, br, bl,
+             name = name, cl = "morris_board_inner", gp = gp,
+             vp = viewport(width=6/7, height=6/7))
+}